requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.0
Pillow>=10.0.0
django-allauth>=0.59.0
requests>=2.31.0
cryptography>=41.0.0
